updated code and aded Service

+++ src/main/java/StudentManagment/service/impl/StudentServiceImpl.java
@@ -3,8 +3,10 @@ package StudentManagment.service.impl;
 import StudentManagment.dto.req.StudentRequestDto;
 import StudentManagment.dto.res.StudentResponseDto;
 import StudentManagment.entity.Student;
+import StudentManagment.entity.User;
 import StudentManagment.mapper.StudentMapper;
 import StudentManagment.repository.StudentRepository;
+import StudentManagment.repository.UserRepository;
 import StudentManagment.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,39 +22,57 @@ public class StudentServiceImpl implements StudentService {
     private StudentRepository repository;
     @Autowired
     private StudentMapper mapper;
+    @Autowired
+    private UserRepository userRepository;
 
 
     @Override
     public List<StudentResponseDto> get() {
-        return repository.findAll().stream().map(mapper::toDto).collect(Collectors.toList());
+        return repository.findAll().stream().map(mapper::toDto)
+                .collect(Collectors.toList());
     }
 
     @Override
     public StudentResponseDto getById(UUID id) {
-        return repository.findById(id).map(mapper::toDto).orElseThrow(() -> new RuntimeException("Student not found"));
+        return repository.findById(id).map(mapper::toDto)
+                .orElseThrow(() -> new RuntimeException("Student not found"));
     }
 
     @Override
     public StudentResponseDto created(StudentRequestDto dto) {
+        User user = userRepository.findById(dto.userId())
+                .orElseThrow(() -> new RuntimeException("User nor found with id: " + dto.userId()));
+
         Student student = mapper.toEntity(dto);
-        Student save = repository.save(student);
-        return mapper.toDto(save);
+        student.setUser(user);
+        repository.save(student);
+        return mapper.toDto(student);
     }
 
     @Override
     public StudentResponseDto updated(UUID id, StudentRequestDto dto) {
-        Student student = repository.findById(id).orElseThrow(() -> new RuntimeException("Student not found"));
+        Student student = repository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Student not found with id" + id));
+        student.setStudentId(dto.studentId());
+        student.setEnrollmentDate(dto.enrollmentDate());
+        student.setGpa(dto.gpa());
+        student.setMajor(dto.major());
+        student.setYear(dto.year());
 
-        student.setName(dto.name());
-        student.setPhone(dto.phone());
-        student.setEmail(dto.email());
-
-        Student updated = repository.save(student);
-        return mapper.toDto(updated);
+        if (dto.userId() != null) {
+            User user = userRepository.findById(id)
+                    .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
+            student.setUser(user);
+        }
+        repository.save(student);
+        return mapper.toDto(student);
     }
 
     @Override
     public void deleted(UUID id) {
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Student not found with id: " + id);
+        }
         repository.deleteById(id);
     }
 }
